refactor: HttpRequest 객체와 요청파일의 절대 경로를 바탕으로 HttpResponse 객체를 생성하는 HttpAssembler 클래스 생성
