+++ src/main/java/webserver/HttpResponse.java
@@ -1,33 +1,20 @@
 package webserver;
 
-import model.RequestType;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.DataOutputStream;
-import java.io.IOException;
-
 public class HttpResponse {
 
-    private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
-    public static void sendResponse200(DataOutputStream dos, byte[] body, HttpRequest httpRequest) {
-        try {
-            dos.writeBytes(response200Header(body.length, httpRequest.getRequestType()));
+    private final String headers;
+    private final byte[] messageBody;
 
-            dos.write(body, 0, body.length);
-            dos.flush();
-        } catch (IOException e) {
-            logger.error(e.getMessage());
-        }
+    public HttpResponse(String headers, byte[] messageBody) {
+        this.headers = headers;
+        this.messageBody = messageBody;
     }
 
-    private static String response200Header(int lengthOfBodyContent, RequestType requestType) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("HTTP/1.1 200 OK \r\n");
-        sb.append("Content-Type: " + requestType.getContentType() + "\r\n");
-        sb.append("Content-Length: " + lengthOfBodyContent + "\r\n");
-        sb.append("\r\n");
+    public String getHeaders() {
+        return headers;
+    }
 
-        return sb.toString();
+    public byte[] getMessageBody() {
+        return messageBody;
     }
 }
